move marker memo struct to types
